diffrule API

diffrule API

Many differentiation methods rely on the notion of "primitive" definitions that comprise "building blocks" that can be composed via various formulations of the chain rule.

To facilitate the definition and use of such rules, DiffBase provides the diffrule API. These methods allows you to define derivative rules, query which rules exist, and symbolically apply the rules to function call expressions.

The complete list of differentiation rules defined by DiffBase can be found in DiffBase/src/rules.jl.

DiffBase.@define_diffrule — Macro.

@define_diffrule f(x) = :(df_dx($x))
@define_diffrule f(x, y) = :(df_dx($x, $y)), :(df_dy($x, $y))
⋮

Define a new differentiation rule for the function f and the given arguments, which should be treated as bindings to Julia expressions.

The RHS should be a function call with a non-splatted argument list, and the LHS should be the derivative expression, or in the n-ary case, an n-tuple of expressions where the ith expression is the derivative of f w.r.t the ith argument. Arguments should be interpolated wherever they are used on the RHS.

Note that differentiation rules are purely symbolic, so no type annotations should be used.

Examples:

@define_diffrule cos(x)          = :(-sin($x))
@define_diffrule /(x, y)         = :(inv($y)), :(-$x / ($y^2))
@define_diffrule polygamma(m, x) = :NaN,       :(polygamma($m + 1, $x))

source

DiffBase.diffrule — Function.

diffrule(f::Symbol, args...)

Return the derivative expression for f at the given argument(s), with the argument(s) interpolated into the returned expression.

In the n-ary case, an n-tuple of expressions will be returned where the ith expression is the derivative of f w.r.t the ith argument.

Examples:

julia> DiffBase.diffrule(:sin, 1)
:(cos(1))

julia> DiffBase.diffrule(:sin, :x)
:(cos(x))

julia> DiffBase.diffrule(:sin, :(x * y^2))
:(cos(x * y ^ 2))

julia> DiffBase.diffrule(:^, :(x + 2), :c)
(:(c * (x + 2) ^ (c - 1)), :((x + 2) ^ c * log(x + 2)))

source

DiffBase.hasdiffrule — Function.

hasdiffrule(f::Symbol, arity::Int)

Return true if a differentiation rule is defined for f and arity, or return false otherwise.

Here, arity refers to the number of arguments accepted by f.

Examples:

julia> DiffBase.hasdiffrule(:sin, 1)
true

julia> DiffBase.hasdiffrule(:sin, 2)
false

julia> DiffBase.hasdiffrule(:-, 1)
true

julia> DiffBase.hasdiffrule(:-, 2)
true

julia> DiffBase.hasdiffrule(:-, 3)
false

source